Move user profile permissions into separate component
